# swagger.json
{"openapi":"3.0.0","paths":{"/work":{"post":{"operationId":"WorkController_createWork","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"properties":{"file":{"type":"string","format":"binary"},"name":{"type":"string"},"chapter":{"type":"number"},"url":{"type":"string"},"category":{"type":"string","enum":["ANIME","MANGA"]}},"$ref":"#/components/schemas/"}}}},"responses":{"201":{"description":""}},"tags":["work"]}},"/work/{id}":{"delete":{"operationId":"WorkController_deleteWork","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["work"]}},"/work/find/{id}":{"get":{"operationId":"WorkController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkHttp"}}}}},"tags":["work"]}},"/work/{id}/update-chapter":{"patch":{"operationId":"WorkController_updateChapter","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChapterDto"}}}},"responses":{"200":{"description":""}},"tags":["work"]}},"/work/{id}/mark-read":{"patch":{"operationId":"WorkController_markRead","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["work"]}},"/work/{id}/mark-unread":{"patch":{"operationId":"WorkController_markUnread","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarkWorkUnreadDto"}}}},"responses":{"200":{"description":""}},"tags":["work"]}},"/work/sync-to-notion ":{"post":{"operationId":"WorkController_syncToNotion","parameters":[],"responses":{"201":{"description":""}},"tags":["work"]}},"/work/fetch-for-workers-read":{"get":{"operationId":"WorkController_fetchForWorkersRead","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkHttp"}}}}}},"tags":["work"]}},"/work/fetch-for-workers-unread":{"get":{"operationId":"WorkController_fetchForWorkersUnread","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkHttp"}}}}}},"tags":["work"]}},"/work/refresh-chapters":{"post":{"operationId":"WorkController_refreshChapters","parameters":[],"responses":{"201":{"description":""}},"tags":["work"]}},"/work/update-work/{id}":{"put":{"operationId":"WorkController_updateWork","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkDto"}}}},"responses":{"200":{"description":""}},"tags":["work"]}},"/work/mark-finished/{id}":{"patch":{"operationId":"WorkController_markFinished","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["work"]}},"/work/upload-work-image/{id}":{"post":{"operationId":"WorkController_uploadWorkImage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"properties":{"file":{"type":"string","format":"binary"}},"$ref":"#/components/schemas/"}}}},"responses":{"201":{"description":""}},"tags":["work"]}},"/work/replace-image-from-notion":{"get":{"operationId":"WorkController_setWorkImageFromNotion","parameters":[],"responses":{"200":{"description":""}},"tags":["work"]}},"/work/fetch-for-works-scraping-report":{"get":{"operationId":"WorkController_fetchForWorksScrapingReportPaginated","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"filter","required":false,"in":"query","schema":{"$ref":"#/components/schemas/RefreshStatus"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkHttp"}}}}}},"tags":["work"]}},"/work/scrapping-report":{"post":{"operationId":"WorkController_scrappingFallback","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScrappingReportDto"}}}},"responses":{"201":{"description":""}},"tags":["work"]}},"/work/dropped/{workId}":{"patch":{"operationId":"WorkController_markWorkAsDropped","parameters":[{"name":"workId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["work"]}},"/work/sync-work":{"post":{"operationId":"WorkController_syncWork","parameters":[],"responses":{"201":{"description":""}},"tags":["work"]}},"/work/list":{"get":{"operationId":"WorkController_listUserWorks","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkHttp"}}}}}},"tags":["work"]}},"/auth/login":{"post":{"operationId":"AuthController_makeSession","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MakeSessionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenModel"}}}}},"tags":["auth"]}},"/auth/user/avatar/upload":{"post":{"operationId":"AuthController_uploadAvatarImage","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"file","$ref":"#/components/schemas/"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/user/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserHttp"}}}}},"tags":["auth"]}},"/auth/access-token":{"post":{"operationId":"AuthController_createAccessToken","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessToken"}}}}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/admin-hash-code":{"post":{"operationId":"AuthController_createAdminHashCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdminHashCodeDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login-mobile":{"post":{"operationId":"AuthController_createMobileSession","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MakeSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/notion/update-database-id":{"post":{"operationId":"AuthController_updateNotionDatabaseId","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNotionDatabaseIdDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/user/analytics":{"get":{"operationId":"AuthController_fetchUserAnalytics","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/user/trial-quote":{"get":{"operationId":"AuthController_getUserTrialQuoteGet","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/user/telegram-status":{"get":{"operationId":"AuthController_getTelegramStatus","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/notification/push/browser/subscribe":{"post":{"operationId":"NotificationController_registerBrowserSubscriber","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscribeUserBrowserNotificationDto"}}}},"responses":{"201":{"description":""}},"tags":["notification"]}},"/notification/push/browser/public-key":{"get":{"operationId":"NotificationController_getPublicKey","parameters":[],"responses":{"200":{"description":""}},"tags":["notification"]}},"/notification/push/telegram/subscribe":{"post":{"operationId":"NotificationController_subscribeInTelegram","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterTelegramChatIdDto"}}}},"responses":{"201":{"description":""}},"tags":["notification"]}},"/notification/push/mobile/subscribe":{"post":{"operationId":"NotificationController_subscribeInMobile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterMobilePushSubscriberDto"}}}},"responses":{"201":{"description":""}},"tags":["notification"]}},"/payment/checkout":{"post":{"operationId":"PaymentController_createIntentPayment","parameters":[],"responses":{"201":{"description":""}},"tags":["payment"]}},"/payment/webhook":{"post":{"operationId":"PaymentController_webhook","parameters":[],"responses":{"201":{"description":""}},"tags":["payment"]}},"/health":{"get":{"operationId":"LoggerController_healthCheck","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}}}},"/metrics":{"get":{"operationId":"PrometheusController_getMetrics","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"Okami API","description":"The Okami rest api","version":"1.0","contact":{}},"tags":[{"name":"okami","description":""}],"servers":[],"components":{"schemas":{"WorkHttp":{"type":"object","properties":{"chapter":{"type":"number"},"hasNewChapter":{"type":"boolean"},"id":{"type":"string"},"imageId":{"type":"string"},"imageUrl":{"type":"string","nullable":true},"isFinished":{"type":"boolean"},"name":{"type":"string"},"url":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"category":{"type":"string"},"isDropped":{"type":"boolean"},"nextChapterUpdatedAt":{"format":"date-time","type":"string","nullable":true},"nextChapter":{"type":"number","nullable":true},"refreshStatus":{"type":"string"}},"required":["chapter","hasNewChapter","id","imageId","imageUrl","isFinished","name","url","updatedAt","category","isDropped","nextChapterUpdatedAt","nextChapter","refreshStatus"]},"UpdateChapterDto":{"type":"object","properties":{"chapter":{"type":"number"}},"required":["chapter"]},"MarkWorkUnreadDto":{"type":"object","properties":{}},"UpdateWorkDto":{"type":"object","properties":{"chapter":{"type":"number"},"url":{"type":"string"},"name":{"type":"string"}},"required":["chapter","url","name"]},"RefreshStatus":{"type":"string","enum":["SUCCESS","FAILED","PENDING"]},"ScrappingReportDto":{"type":"object","properties":{"status":{"type":"string"},"workId":{"type":"string"}},"required":["status","workId"]},"MakeSessionDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}},"required":["email","password"]},"TokenModel":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"UserHttp":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"avatarImageUrl":{"type":"object"},"avatarImageId":{"type":"string"},"finishedWorksCount":{"type":"number"},"readingWorksCount":{"type":"number"},"notionDatabaseId":{"type":"string"}},"required":["id","name","email","avatarImageUrl","avatarImageId","finishedWorksCount","readingWorksCount"]},"AccessToken":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"CreateUserDto":{"type":"object","properties":{}},"CreateAdminHashCodeDto":{"type":"object","properties":{"hashCodeKey":{"type":"string"}},"required":["hashCodeKey"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string"},"newPassword":{"type":"string"},"adminHashCode":{"type":"string"}},"required":["email","newPassword","adminHashCode"]},"UpdateNotionDatabaseIdDto":{"type":"object","properties":{"notionDatabaseId":{"type":"string"}},"required":["notionDatabaseId"]},"SubscribeUserBrowserNotificationDto":{"type":"object","properties":{}},"RegisterTelegramChatIdDto":{"type":"object","properties":{}},"RegisterMobilePushSubscriberDto":{"type":"object","properties":{}}}}}